requirements.txt
appdirs==1.4.0
backports-abc==0.5
beautifulsoup4==4.5.3
bokeh==0.12.4
bottery
bs4==0.0.1
decorator==4.0.11
ipdb==0.10.2
ipython==5.2.1
ipython-genutils==0.1.0
Jinja2==2.9.5
MarkupSafe==0.23
mongoengine
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pymongo==3.4.0
pyparsing==2.1.10
pyTelegramBotAPI==2.3.0
python-dateutil==2.6.0
python-decouple
pytz==2016.10
PyYAML==3.12
requests==2.13.0
simplegeneric==0.8.1
six==1.10.0
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
